Tests moved to individual project
