+++ cs/HomeExercisesTests/HomeExercisesTests.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net48</TargetFramework>
+        <LangVersion>8</LangVersion>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="NUnit" Version="3.12.0" />
+      <PackageReference Include="NUnit.Engine" Version="3.12.0-beta1" />
+      <PackageReference Include="NUnit3TestAdapter" Version="4.0.0-alpha.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\HomeExercises\HomeExercises.csproj" />
+    </ItemGroup>
+
+</Project>
