+++ cs/HomeExercisesTests/NumberValidatorTests.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
+using HomeExercises;
 
-namespace HomeExercises
+namespace HomeExercisesTests
 {
 	public class NumberValidatorTests
 	{
